fix duplication by moving to core
